+++ crates/distribution-types/src/file.rs
@@ -7,7 +7,7 @@ use thiserror::Error;
 use pep440_rs::{VersionSpecifiers, VersionSpecifiersParseError};
 use pypi_types::{DistInfoMetadata, Hashes, Yanked};
 use url::Url;
-use uv_auth::safe_copy_url_auth_to_str;
+use uv_auth::GLOBAL_AUTH_STORE;
 
 /// Error converting [`pypi_types::File`] to [`distribution_type::File`].
 #[derive(Debug, Error)]
@@ -53,13 +53,13 @@ impl File {
             size: file.size,
             upload_time_utc_ms: file.upload_time.map(|dt| dt.timestamp_millis()),
             url: if file.url.contains("://") {
-                let url = safe_copy_url_auth_to_str(base, &file.url)
-                    .map_err(|err| FileConversionError::Url(file.url.clone(), err))?
-                    .map(|url| url.to_string())
-                    .unwrap_or(file.url);
-
-                FileLocation::AbsoluteUrl(url)
+                // Copy over any credentials from the global store.
+                let url = Url::parse(&file.url)
+                    .map_err(|err| FileConversionError::Url(file.url.clone(), err))?;
+                let url = GLOBAL_AUTH_STORE.with_url_encoded_auth(url);
+                FileLocation::AbsoluteUrl(url.to_string())
             } else {
+                // It's assumed that the base URL already contains any necessary credentials.
                 FileLocation::RelativeUrl(base.to_string(), file.url)
             },
             yanked: file.yanked,