+++ crates/uv-distribution/src/distribution_database.rs
@@ -472,6 +472,7 @@ impl<'a, Context: BuildContext + Send + Sync> DistributionDatabase<'a, Context> 
                 let archive = self
                     .cache
                     .persist(temp_dir.into_path(), wheel_entry.path())
+                    .await
                     .map_err(Error::CacheRead)?;
                 Ok(archive)
             }
@@ -553,6 +554,7 @@ impl<'a, Context: BuildContext + Send + Sync> DistributionDatabase<'a, Context> 
                 let archive = self
                     .cache
                     .persist(temp_dir.into_path(), wheel_entry.path())
+                    .await
                     .map_err(Error::CacheRead)?;
                 Ok(archive)
             }