+++ crates/uv/tests/pip_install.rs
@@ -1419,11 +1419,11 @@ fn install_upgrade() {
 #[test]
 fn install_constraints_txt() -> Result<()> {
     let context = TestContext::new("3.12");
-    let requirementstxt = context.temp_dir.child("requirements.txt");
-    requirementstxt.write_str("django==5.0b1")?;
+    let requirements_txt = context.temp_dir.child("requirements.txt");
+    requirements_txt.write_str("anyio==3.7.0")?;
 
     let constraints_txt = context.temp_dir.child("constraints.txt");
-    constraints_txt.write_str("sqlparse<0.4.4")?;
+    constraints_txt.write_str("idna<3.4")?;
 
     uv_snapshot!(command(&context)
             .arg("-r")
@@ -1438,9 +1438,9 @@ fn install_constraints_txt() -> Result<()> {
     Resolved 3 packages in [TIME]
     Downloaded 3 packages in [TIME]
     Installed 3 packages in [TIME]
-     + asgiref==3.7.2
-     + django==5.0b1
-     + sqlparse==0.4.3
+     + anyio==3.7.0
+     + idna==3.3
+     + sniffio==1.3.0
     "###
     );
 
@@ -1452,10 +1452,10 @@ fn install_constraints_txt() -> Result<()> {
 fn install_constraints_inline() -> Result<()> {
     let context = TestContext::new("3.12");
     let requirementstxt = context.temp_dir.child("requirements.txt");
-    requirementstxt.write_str("django==5.0b1\n-c constraints.txt")?;
+    requirementstxt.write_str("anyio==3.7.0\n-c constraints.txt")?;
 
     let constraints_txt = context.temp_dir.child("constraints.txt");
-    constraints_txt.write_str("sqlparse<0.4.4")?;
+    constraints_txt.write_str("idna<3.4")?;
 
     uv_snapshot!(command(&context)
             .arg("-r")
@@ -1468,9 +1468,9 @@ fn install_constraints_inline() -> Result<()> {
     Resolved 3 packages in [TIME]
     Downloaded 3 packages in [TIME]
     Installed 3 packages in [TIME]
-     + asgiref==3.7.2
-     + django==5.0b1
-     + sqlparse==0.4.3
+     + anyio==3.7.0
+     + idna==3.3
+     + sniffio==1.3.0
     "###
     );
 
@@ -2717,6 +2717,44 @@ fn dry_run_install_then_upgrade() -> std::result::Result<(), Box<dyn std::error:
     Would install 1 package
      - httpx==0.25.0
      + httpx==0.25.1
+    "###
+    );
+
+    Ok(())
+}
+
+/// Raise an error when a direct URL's `Requires-Python` constraint is not met.
+#[test]
+fn requires_python_direct_url() -> Result<()> {
+    let context = TestContext::new("3.12");
+
+    // Create an editable package with a `Requires-Python` constraint that is not met.
+    let editable_dir = assert_fs::TempDir::new()?;
+    let pyproject_toml = editable_dir.child("pyproject.toml");
+    pyproject_toml.write_str(
+        r#"[project]
+name = "example"
+version = "0.0.0"
+dependencies = [
+  "anyio==4.0.0"
+]
+requires-python = "<=3.8"
+"#,
+    )?;
+
+    uv_snapshot!(command(&context)
+        .arg(format!("example @ {}", editable_dir.path().display())), @r###"
+    success: false
+    exit_code: 1
+    ----- stdout -----
+
+    ----- stderr -----
+      × No solution found when resolving dependencies:
+      ╰─▶ Because the current Python version (3.12.1) does not satisfy Python<=3.8
+          and example==0.0.0 depends on Python<=3.8, we can conclude that
+          example==0.0.0 cannot be used.
+          And because only example==0.0.0 is available and you require example, we
+          can conclude that the requirements are unsatisfiable.
     "###
     );